dépot colis : wip

+++ www/SQL/baseCreation.sql
@@ -104,7 +104,7 @@ CREATE TABLE DeliveryType(
 
 CREATE TABLE Parcel(
 	id INT NOT NULL AUTO_INCREMENT,
-	weight INT NOT NULL,
+	weight FLOAT NOT NULL,
 	status_id INT NOT NULL,
 	is_deleted INT DEFAULT 0,
 	delivery_type INT NOT NULL,
